move RolloutStorage to utils
